# swagger/SocialnetApiDocs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Socialnet API",
    "description": "Code Lounge",
    "version": "1.6"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/load": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Загрузка геоданных",
        "operationId": "loadGeoData",
        "responses": {
          "200": {
            "description": "Геоданные успешно загружены",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/friends/{id}/approve": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Одобрение запроса в друзья",
        "operationId": "approveFriendRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта для одобрения запроса в друзья",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Запрос в друзья успешно одобрен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FriendShortDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/friends/unblock/{id}": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Разблокировка аккаунта",
        "operationId": "unblockAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта для разблокировки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Аккаунт успешно разблокирован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FriendShortDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/friends/block/{id}": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Блокировка аккаунта",
        "operationId": "blockAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта для блокировки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Аккаунт успешно заблокирован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FriendShortDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/friends/{id}/request": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Отправка запроса в друзья",
        "operationId": "makeFriendRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта для запроса в друзья",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Запрос в друзья успешно отправлен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FriendShortDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/friends/subscribe/{id}": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Подписка на аккаунт",
        "operationId": "getSubscription",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта для подписки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Подписка успешно оформлена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FriendShortDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Регистрация пользователя",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Пользователь успешно зарегистрирован",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Обновление токена",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Токен успешно обновлен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/auth/password/recovery/{linkId}": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Установка нового пароля",
        "operationId": "setNewPassword",
        "parameters": [
          {
            "name": "linkId",
            "in": "path",
            "description": "Токен для восстановления пароля",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Новый пароль успешно установлен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/auth/password/recovery/": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Восстановление пароля",
        "operationId": "recoverPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Запрос на восстановление пароля успешно отправлен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Выход из системы",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Успешный выход",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Вход в систему",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/auth/change-password-link": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Изменение пароля",
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Запрос на изменение пароля успешно отправлен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/auth/change-email-link": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Изменение Email",
        "operationId": "changeEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Запрос на изменение Email успешно отправлен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/geo/country": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Получение списка стран",
        "operationId": "getCountries",
        "responses": {
          "200": {
            "description": "Список стран успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/geo/country/{countryId}/city": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Получение списка городов по ID страны",
        "operationId": "getCitiesByCountryId",
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "description": "ID страны для получения списка городов",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список городов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/friends": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Получение друзей по параметрам",
        "operationId": "getFriendsByParameters",
        "parameters": [
          {
            "name": "friendSearchDto",
            "in": "query",
            "description": "Параметры поиска друзей",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FriendSearchDto"
            }
          },
          {
            "name": "pageDto",
            "in": "query",
            "description": "Данные страницы поисковой выдачи",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Друзья успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/friends/recommendations": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Получение рекомендаций друзей",
        "operationId": "getFriendsRecommendations",
        "parameters": [
          {
            "name": "friendSearchDto",
            "in": "query",
            "description": "Параметры поиска друзей",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FriendSearchDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Рекомендации друзей успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendShortDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/friends/count": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Получение количества входящих запросов в друзья",
        "operationId": "getIncomingFriendRequests",
        "responses": {
          "200": {
            "description": "Успешное получение количества запросов",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/auth/captcha": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Получение капчи",
        "operationId": "getCaptcha",
        "responses": {
          "200": {
            "description": "Капча успешно получена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CaptchaDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/friends/{id}": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "Удаление отношений",
        "operationId": "deleteRelationship",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта для удаления отношений",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Отношения успешно удалены"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FriendShortDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID",
            "format": "uuid"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Флаг удаления"
          },
          "statusCode": {
            "type": "string",
            "description": "Код статуса",
            "enum": [
              "FRIEND",
              "REQUEST_TO",
              "REQUEST_FROM",
              "BLOCKED",
              "DECLINED",
              "SUBSCRIBED",
              "NONE",
              "WATCHING",
              "REJECTING",
              "RECOMMENDATION"
            ]
          },
          "friendId": {
            "type": "string",
            "description": "ID субъекта",
            "format": "uuid"
          },
          "idFriend": {
            "type": "string",
            "description": "ID объекта",
            "format": "uuid"
          },
          "previousStatusCode": {
            "type": "string",
            "description": "Код статуса",
            "enum": [
              "FRIEND",
              "REQUEST_TO",
              "REQUEST_FROM",
              "BLOCKED",
              "DECLINED",
              "SUBSCRIBED",
              "NONE",
              "WATCHING",
              "REJECTING",
              "RECOMMENDATION"
            ]
          },
          "rating": {
            "type": "integer",
            "description": "Рейтинг",
            "format": "int32"
          }
        },
        "description": "Краткая информация об отношениях"
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Электронная почта"
          },
          "password1": {
            "type": "string",
            "description": "Пароль"
          },
          "password2": {
            "type": "string",
            "description": "Подтверждение пароля"
          },
          "firstName": {
            "type": "string",
            "description": "Имя"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия"
          },
          "captchaCode": {
            "type": "string",
            "description": "Код капчи"
          },
          "captchaSecret": {
            "type": "string",
            "description": "Секрет капчи"
          }
        },
        "description": "Данные для регистрации"
      },
      "RefreshTokenRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Токен обновления"
          }
        },
        "description": "Запрос на обновление access токена"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Токен доступа"
          },
          "refreshToken": {
            "type": "string",
            "description": "Токен обновления"
          }
        },
        "description": "Форма ответа с jwt"
      },
      "NewPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Новый пароль"
          }
        },
        "description": "Данные для установки нового пароля"
      },
      "EmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "recover@example.com"
          }
        },
        "description": "DTO с Email"
      },
      "LoginForm": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя"
          }
        },
        "description": "Форма для входа"
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "passwordOld": {
            "type": "string",
            "description": "Старый пароль"
          },
          "passwordNew": {
            "type": "string",
            "description": "Новый пароль"
          }
        },
        "description": "DTO для смены пароля"
      },
...
      "SortDto": {
        "type": "object",
        "properties": {
          "unsorted": {
            "type": "boolean",
            "description": "Не отсортировано"
          },
          "sorted": {
            "type": "boolean",
            "description": "Отсортировано"
          },
          "empty": {
            "type": "boolean",
            "description": "Пусто"
          }
        },
        "description": "Данные сортировки"
      },
...
      "CountryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID страны",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "Название страны"
          },
          "cities": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "description": "Список городов"
            }
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "description": "Данные страны"
      },
      "CityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID города",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "Название города"
          },
          "countryId": {
            "type": "integer",
            "description": "ID страны",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "description": "Данные города"
      },
      "FriendSearchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID аккаунта",
            "format": "uuid"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Флаг удаления"
          },
          "idFrom": {
            "type": "integer",
            "description": "ID субъекта отношений",
            "format": "int32"
          },
          "statusCode": {
            "type": "string",
            "description": "Код статуса",
            "enum": [
              "FRIEND",
              "REQUEST_TO",
              "REQUEST_FROM",
              "BLOCKED",
              "DECLINED",
              "SUBSCRIBED",
              "NONE",
              "WATCHING",
              "REJECTING",
              "RECOMMENDATION"
            ]
          },
          "idTo": {
            "type": "integer",
            "description": "ID объекта отношений",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "description": "Имя"
          },
          "birthDateFrom": {
            "type": "string",
            "description": "Дата рождения от",
            "format": "date"
          },
          "birthDateTo": {
            "type": "string",
            "description": "Дата рождения до",
            "format": "date"
          },
          "city": {
            "type": "string",
            "description": "Город"
          },
          "country": {
            "type": "string",
            "description": "Страна"
          },
          "ageFrom": {
            "type": "integer",
            "description": "Возраст от",
            "format": "int32"
          },
          "ageTo": {
            "type": "integer",
            "description": "Возраст до",
            "format": "int32"
          },
          "previousStatusCode": {
            "type": "string",
            "description": "Код статуса",
            "enum": [
              "FRIEND",
              "REQUEST_TO",
              "REQUEST_FROM",
              "BLOCKED",
              "DECLINED",
              "SUBSCRIBED",
              "NONE",
              "WATCHING",
              "REJECTING",
              "RECOMMENDATION"
            ]
          }
        },
        "description": "Форма поиска друзей"
      },
      "PageDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Страница (номер)",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "description": "Размер",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "description": "Сортировка",
            "items": {
              "type": "string",
              "description": "Сортировка"
            }
          }
        },
        "description": "Данные страницы"
      },
      "PageResponseDtoObject": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "description": "Общее количество элементов",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "description": "Общее количество страниц",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "first": {
            "type": "boolean",
            "description": "Первая страница"
          },
          "last": {
            "type": "boolean",
            "description": "Последняя страница"
          },
          "numberOfElements": {
            "type": "integer",
            "description": "Количество элементов на странице",
            "format": "int32"
          },
          "empty": {
            "type": "boolean",
            "description": "Пустая страница"
          },
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableDto"
          },
          "content": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "description": "Содержимое страницы",
            "items": {
              "type": "object"
            }
          }
        },
        "description": "Ответ с данными страницы"
      },
      "CountDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "Количество",
            "format": "int32"
          }
        },
        "description": "Счетчик"
      },
...
      "CaptchaDto": {
        "type": "object",
        "properties": {
          "secret": {
            "type": "string",
            "description": "Секретный ключ капчи"
          },
          "image": {
            "type": "string",
            "description": "Изображение капчи с заголовком"
          }
        },
        "description": "Информация о капче"
      },
...
    }
  }
}